pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "postmule"
version = "0.1.0"
description = "AI-powered physical mail management — the open source PayTrust replacement"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["mail", "virtual mailbox", "bills", "OCR", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "pdfplumber>=0.11",
    "pytesseract>=0.3",
    "Pillow>=10",
    "google-auth>=2",
    "google-auth-oauthlib>=1",
    "google-api-python-client>=2",
    "google-generativeai>=0.7",
    "cryptography>=42",
    "keyring>=25",
    "flask>=3",
    "pyyaml>=6",
    "requests>=2",
    "rapidfuzz>=3",
    "click>=8",
    "playwright>=1.44",
    "jinja2>=3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-mock>=3",
    "black>=24",
    "ruff>=0.4",
]

[project.scripts]
postmule = "postmule.cli:main"

[project.urls]
Homepage = "https://github.com/PostMule/postmule"
Repository = "https://github.com/PostMule/postmule"
Issues = "https://github.com/PostMule/postmule/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=postmule --cov-report=term-missing"
# TODO: raise --cov-fail-under back to 80 once stub files are implemented

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
